+++ CrystalSpace/CrystalSpace/MarchingCubesAlgo.h
@@ -15,9 +15,9 @@ class MarchingCubesAlgo
 {
 public:
 
-	size_t build(const Shape::Volume<double>& volume, const double isoLevel);
+	size_t build(const Shape::Volume<float>& volume, const double isoLevel);
 
-	void build(const SparseVolume<double>& volume, const float isoLevel);
+	void build(const SparseVolume<float>& volume, const float isoLevel);
 
 	int march(const MCCell& cell, const double isoLevel);
